+++ src/CosmoSim/__init__.py
@@ -12,7 +12,7 @@ import os, sys
 
 import traceback
 
-__version__ = "2.7.0"
+__version__ = "2.7.1"
 
 ModelSpec = cs.ModelSpec
 SourceSpec = cs.SourceSpec